bot/requirements.txt
pyTelegramBotAPI==4.12.0
requests
python-dotenv
